Created HTML Conversion Function
This function will convert HTML with HTML-safe characters back to normal HTML so it can be displayed as such. This is very useful because the special characters have to be converted to HTML-safe so that they don't give problems for MySQL when attempting to put these into the database. Since this conversion will happen after the variable has been pulled out of MySQL, it will be safe to use and will display properly.