Created getVar function
The getVar function can be used to safely get the value of any variable as long as the name is unique. It will check to make sure if the variable is a Get, Post or normal variable and confirm that it exists at all to prevent any errors. If it doesn't exist at all, it will set the value to whatever you set as the blank value, which is the second parameter this function accepts. This should save me a lot of time.