Dynamic Buttons Display - not done
There is much more testing needed for this and I will need to make sure what it's actually doing after the button is pressed to confirm there are no issues. Looks like when I pressed a button it actually assigned a value to the resident and moved on to the next one. Need to figure out why it did that.