Phone numbers ready to call and check in of exports
Got the phone numbers ready to call list fully accurate. The territories ready to check in is now filtered on whether or not the exported addresses have been returned. Started work on creating a way to view the exported addresses for a territory by date and publisher. Haven't yet gotten the drill down to see which addresses were exported or the link that actually returns exported addresses.