WIP - not ready
This is still very much in the early stages, but figured out the proper 
methodology on this. I will have to use jquery post command to get this 
working. Tried to get curl to work, but it's fighting me but when I do a 
normal form submit, it does create the JSON data that I need. That means 
that I need to use jQuery to grab that, pull it into a local hidden form 
and then submit it. This will need to be done every time we do a 
download. 

I think that while the token only lasts for 3600 secs, the code and 
other data will last longer. I think we will want to have it save the 
code, redirect_uri, client_id, and client_secret fields all in a table 
as those won't change often. It's a little weird to do, but we may have 
to then run a local batch file to open the page on the site so that the 
jQuery will run because jQuery requires a browser.